Validation, database, users module

+++ src/courses/courses.controller.ts
@@ -1,38 +1,61 @@
-import { Body, Controller, Delete, Get, Param, ParseIntPipe, Post, Put } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  NotFoundException,
+  Post,
+  Put,
+} from '@nestjs/common';
+
 import { CoursesService } from './courses.service';
 import { CourseDto } from './dto/courses.dto';
+import { Id } from '../common/id.decorator';
 
 @Controller('courses')
 export class CoursesController {
-    constructor(private readonly coursesService: CoursesService) {
+  constructor(private readonly coursesService: CoursesService) {}
 
+  @Get(':id')
+  async getOne(@Id() id: number) {
+    const course = await this.coursesService.getOne(id);
+
+    if (!course) {
+      throw new NotFoundException();
     }
 
-    @Get(':id')
-    getOne(@Param('id', ParseIntPipe) id: number) {
-        console.log(typeof id);
-        const course = this.coursesService.getOne(id);
+    return course;
+  }
 
-        return course;
+  @Get()
+  async getAll() {
+    return this.coursesService.getAll();
+  }
+
+  @Post()
+  async create(@Body() data: CourseDto) {
+    return this.coursesService.create(data);
+  }
+
+  @Put(':id')
+  async modify(@Id() id: number, @Body() data: CourseDto) {
+    const course = await this.coursesService.getOne(id);
+
+    if (!course) {
+      throw new NotFoundException();
     }
 
-    @Get()
-    getAll() {
-        return this.coursesService.getAll();
+    return this.coursesService.modify(id, data);
+  }
+
+  @Delete(':id')
+  async delete(@Id() id: number) {
+    const course = await this.coursesService.getOne(id);
+
+    if (!course) {
+      throw new NotFoundException();
     }
 
-    @Post()
-    create(@Body() data: CourseDto) {
-        return this.coursesService.create(data);
-    }
-
-    @Put(':id')
-    modify(@Param('id', ParseIntPipe) id: number, @Body() data: CourseDto) {
-        return this.coursesService.modify(id, data);
-    }
-
-    @Delete(':id')
-    delete(@Param('id', ParseIntPipe) id: number) {
-        this.coursesService.delete(id);
-    }
+    await this.coursesService.delete(id);
+  }
 }
